pyproject.toml
[project]
name = "parallellm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "colorama>=0.4.6",
    "openai>=1.107.2",
    "pillow>=10.0.0",
    "polars>=1.32.3",
]

[tool.setuptools.packages.find]
include = ["parallellm*"]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.1",
]
